Applied changes from code review
