--- src/Directory.Build.props
@@ -1,8 +1,0 @@
-<Project>
-  <ItemGroup Condition="$(MSBuildProjectName.EndsWith('Tests'))">
-    <PackageReference Include="SonarAnalyzer.CSharp" Version="10.27.0.140913">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-</Project>
