+++ src/UltimatePDF_ExternalLogic.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UltimatePDF_ExternalLogic.T
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UltimatePDF_ExternalLogic.Test.Helpers", "UltimatePDF_ExternalLogic.Test.Helpers\UltimatePDF_ExternalLogic.Test.Helpers.csproj", "{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,6 +42,10 @@ Global
 		{15362A99-BC62-456D-AFBD-433475909B31}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{15362A99-BC62-456D-AFBD-433475909B31}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{15362A99-BC62-456D-AFBD-433475909B31}.Release|Any CPU.Build.0 = Release|Any CPU
+		{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -48,6 +54,7 @@ Global
 		{9852B0A7-DDC6-4876-8539-04FFA670B7B3} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{7DE787C3-F07C-4739-9B23-C82F686332AB} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{15362A99-BC62-456D-AFBD-433475909B31} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{74F254C6-91BF-4EE2-ABF3-A5ABC66F9B8D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D72CC5ED-6EF0-48FF-A71D-EB2075E02A04}
