+++ src/UltimatePDF_ExternalLogic/Utils/RestSender.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using System.Net.Http;
 using System.Threading.Tasks;
@@ -5,10 +6,10 @@ using OutSystems.UltimatePDF_ExternalLogic.Management.Troubleshooting;
 using OutSystems.UltimatePDF_ExternalLogic.Structures;
 
 namespace UltimatePDF_ExternalLogic.Utils {
-    internal class RestSender {
+    internal class RestSender : IDisposable {
         private readonly RestCaller restCaller;
         private readonly Logger logger;
-        private readonly HttpClient _client;
+        private readonly HttpClient client;
 
         public RestSender(RestCaller restCaller, Logger logger)
             : this(restCaller, logger, new HttpClientHandler()) { }
@@ -16,7 +17,7 @@ namespace UltimatePDF_ExternalLogic.Utils {
         internal RestSender(RestCaller restCaller, Logger logger, HttpMessageHandler handler) {
             this.restCaller = restCaller;
             this.logger = logger;
-            _client = new HttpClient(handler);
+            client = new HttpClient(handler);
         }
 
         internal async Task RestSendPDFAsync(byte[] pdf) {
@@ -44,8 +45,12 @@ namespace UltimatePDF_ExternalLogic.Utils {
             request.Headers.Add("Authorization", $"Bearer {token}");
             request.Content = new StreamContent(new MemoryStream(binary));
             request.Content.Headers.Add("Content-Type", contentType);
-            using var response = await _client.SendAsync(request);
+            using var response = await client.SendAsync(request);
             response.EnsureSuccessStatusCode();
+        }
+
+        public void Dispose() {
+            client.Dispose();
         }
     }
 }
